backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
pyaudio
pydub
gpiozero
google-generativeai
python-dotenv
